refactor tabs and radio stuff out
